Ignore construction debris

+++ src/input.inl
@@ -123,8 +123,8 @@ static int parse_mouse_event(struct tb_event *event, const char *buf, int len) {
     if ((b & 32) != 0) event->meta |= TB_META_MOTION;
 
     // the coord is 1,1 for upper left
-    event->x = (uint8_t)buf[4] - 1 - 32;
-    event->y = (uint8_t)buf[5] - 1 - 32;
+    event->x = (uint16_t)buf[4] - 1 - 32;
+    event->y = (uint16_t)buf[5] - 1 - 32;
 
     if (event->key > TB_KEY_MOUSE_RELEASE && !(event->meta & TB_META_MOTION)) { // click
       if (is_double_click(event->key, event->x, event->y)) {
